requirements_bare.txt
pytz
pyyaml
click
requests
psycopg2-binary
sqlalchemy
beautifulsoup4
#lxml
#html5lib
regex
